requirements.txt
azure-functions
azure-storage-blob
azure-search-documents
azure-core
requests
